pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prayerhub-mini"
version = "0.1.0"
description = "PrayerHub Mini prototype service"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "PrayerHub" }]
dependencies = [
  "APScheduler==3.10.4",
  "Flask==3.0.3",
  "PyYAML==6.0.2",
  "Werkzeug==3.0.3",
  "requests==2.32.3",
]

[tool.setuptools.packages.find]
where = ["src"]
